deprecate `pubkeys_sigs_of`
